update train code

+++ train_model.py
@@ -67,7 +67,8 @@ def update_policy(
     start_time = time.perf_counter()
     device = get_device_from_parameters(policy)
     policy.train()
-    with torch.autocast(device_type=device.type) if use_amp else nullcontext():
+    # dtype=torch.bfloat16 추가!
+    with torch.autocast(device_type=device.type, dtype=torch.bfloat16) if use_amp else nullcontext():
         loss, output_dict = policy.forward(batch)
         # TODO(rcadene): policy.unnormalize_outputs(out_dict)
     grad_scaler.scale(loss).backward()
